First pass at factoring footer out of headless XF code.
