+++ www/index-main.js
@@ -11,5 +11,6 @@ wmf.imageDimming = require('wikimedia-page-library').DimImagesTransform
 wmf.themes = require('wikimedia-page-library').ThemeTransform
 wmf.platform = require('wikimedia-page-library').PlatformTransform
 wmf.sectionTransformation = require('./js/sectionTransformation')
+wmf.footerTransformation = require('./js/footerTransformation')
 
 window.wmf = wmf
