+++ www/js/sectionTransformation.js
@@ -2,10 +2,6 @@
 const requirements = {
   editButtons: require('./transforms/addEditButtons'),
   utilities: require('./utilities'),
-  footerReadMore: require('wikimedia-page-library').FooterReadMore,
-  footerMenu: require('wikimedia-page-library').FooterMenu,
-  footerLegal: require('wikimedia-page-library').FooterLegal,
-  footerContainer: require('wikimedia-page-library').FooterContainer,
   filePages: require('./transforms/disableFilePageEdit'),
   tables: require('./transforms/collapseTables'),
   themes: require('wikimedia-page-library').ThemeTransform,
@@ -29,14 +25,13 @@ class Language {
 }
 
 class Article {
-  constructor(ismain, title, displayTitle, description, editable, language, hasReadMore) {
+  constructor(ismain, title, displayTitle, description, editable, language) {
     this.ismain = ismain
     this.title = title
     this.displayTitle = displayTitle
     this.description = description
     this.editable = editable
     this.language = language
-    this.hasReadMore = hasReadMore
   }
   descriptionParagraph() {
     if(this.description !== undefined && this.description.length > 0){
@@ -108,86 +103,6 @@ class Section {
   }
 }
 
-class Footer {
-  constructor(article, menuItems, readMoreItemCount, localizedStrings, proxyURL) {
-    this.article = article
-    this.menuItems = menuItems
-    this.readMoreItemCount = readMoreItemCount
-    this.localizedStrings = localizedStrings
-    this.proxyURL = proxyURL
-  }
-  addContainer() {
-    if (requirements.footerContainer.isContainerAttached(document) === false) {
-      document.querySelector('body').appendChild(requirements.footerContainer.containerFragment(document))
-      window.webkit.messageHandlers.footerContainerAdded.postMessage('added')
-    }
-  }
-  addDynamicBottomPadding() {
-    window.addEventListener('resize', function(){requirements.footerContainer.updateBottomPaddingToAllowReadMoreToScrollToTop(window)})
-  }
-  addMenu() {
-    requirements.footerMenu.setHeading(this.localizedStrings.menuHeading, 'pagelib_footer_container_menu_heading', document)
-    this.menuItems.forEach(item => {
-      let title = ''
-      let subtitle = ''
-      let menuItemTypeString = ''
-      switch(item) {
-      case requirements.footerMenu.MenuItemType.languages:
-        menuItemTypeString = 'languages'
-        title = this.localizedStrings.menuLanguagesTitle
-        break
-      case requirements.footerMenu.MenuItemType.lastEdited:
-        menuItemTypeString = 'lastEdited'
-        title = this.localizedStrings.menuLastEditedTitle
-        subtitle = this.localizedStrings.menuLastEditedSubtitle
-        break
-      case requirements.footerMenu.MenuItemType.pageIssues:
-        menuItemTypeString = 'pageIssues'
-        title = this.localizedStrings.menuPageIssuesTitle
-        break
-      case requirements.footerMenu.MenuItemType.disambiguation:
-        menuItemTypeString = 'disambiguation'
-        title = this.localizedStrings.menuDisambiguationTitle
-        break
-      case requirements.footerMenu.MenuItemType.coordinate:
-        menuItemTypeString = 'coordinate'
-        title = this.localizedStrings.menuCoordinateTitle
-        break
-      case requirements.footerMenu.MenuItemType.talkPage:
-        menuItemTypeString = 'talkPage'
-        title = this.localizedStrings.menuTalkPageTitle
-        break
-      default:
-      }
-      const itemSelectionHandler = payload => window.webkit.messageHandlers.footerMenuItemClicked.postMessage({'selection': menuItemTypeString, 'payload': payload})
-      requirements.footerMenu.maybeAddItem(title, subtitle, item, 'pagelib_footer_container_menu_items', itemSelectionHandler, document)
-    })
-  }
-  addReadMore() {
-    if (this.article.hasReadMore){
-      requirements.footerReadMore.setHeading(this.localizedStrings.readMoreHeading, 'pagelib_footer_container_readmore_heading', document)
-      const saveButtonTapHandler = title => window.webkit.messageHandlers.footerReadMoreSaveClicked.postMessage({'title': title})
-      const titlesShownHandler = titles => {
-        window.webkit.messageHandlers.footerReadMoreTitlesShown.postMessage(titles)
-        requirements.footerContainer.updateBottomPaddingToAllowReadMoreToScrollToTop(window)
-      }
-      requirements.footerReadMore.add(this.article.title, this.readMoreItemCount, 'pagelib_footer_container_readmore_pages', this.proxyURL, saveButtonTapHandler, titlesShownHandler, document)
-    }
-  }
-  addLegal() {
-    const licenseLinkClickHandler = () => window.webkit.messageHandlers.footerLegalLicenseLinkClicked.postMessage('linkClicked')
-    const viewInBrowserLinkClickHandler = () => window.webkit.messageHandlers.footerBrowserLinkClicked.postMessage('linkClicked')
-    requirements.footerLegal.add(document, this.localizedStrings.licenseString, this.localizedStrings.licenseSubstitutionString, 'pagelib_footer_container_legal', licenseLinkClickHandler, this.localizedStrings.viewInBrowserString, viewInBrowserLinkClickHandler)
-  }
-  add() {
-    this.addContainer()
-    this.addDynamicBottomPadding()
-    this.addMenu()
-    this.addReadMore()
-    this.addLegal()
-  }
-}
-
 const processResponseStatus = response => {
   if (response.status === 200) { // can use status 0 if loading local files
     return Promise.resolve(response)
@@ -238,9 +153,7 @@ const performEarlyNonSectionTransforms = article => {
 }
 
 //late so they won't delay section fragment processing
-const performLateNonSectionTransforms = (article, proxyURL) => {
-  const footer = new Footer(article, this.menuItems, 3, this.localizedStrings, proxyURL)
-  footer.add()
+const performLateNonSectionTransforms = article => {
   // 'themes.classifyElements()' needs to happen once after body elements are present. it
   // classifies some tricky elements like math formula images (see 'enwiki > Quadradic formula')
   requirements.themes.classifyElements(document)
@@ -264,7 +177,7 @@ const scrollToSection = hash => {
   }
 }
 
-const fetchTransformAndAppendSectionsToDocument = (article, proxyURL, articleSectionsURL, hash) => {
+const fetchTransformAndAppendSectionsToDocument = (article, articleSectionsURL, hash, successCallback) => {
   performEarlyNonSectionTransforms(article)
   const mainContentDiv = document.querySelector('div.content')
   fetch(articleSectionsURL)
@@ -272,8 +185,9 @@ const fetchTransformAndAppendSectionsToDocument = (article, proxyURL, articleSec
   .then(extractResponseJSON)
   .then(extractJSONSections)
   .then(sections => transformAndAppendSectionsToMainContentDiv(sections, article, mainContentDiv))
-  .then(() => performLateNonSectionTransforms(article, proxyURL))
+  .then(() => performLateNonSectionTransforms(article))
   .then(() => scrollToSection(hash))
+  .then(successCallback)
   .catch(error => console.log(`Promise was rejected with error: ${error}`))
 }
 
@@ -283,4 +197,3 @@ exports.localizedStrings = undefined
 exports.fetchTransformAndAppendSectionsToDocument = fetchTransformAndAppendSectionsToDocument
 exports.Language = Language
 exports.Article = Article
-exports.menuItems = undefined
